+++ noodles-vcf/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## Unreleased
+
+### Changed
+
+  * vcf/io/writer/record/reference_bases: Resolve IUPAC ambiguity codes
+    ([#268]).
+
+    Any occurrance of an IUPAC ambiguity code is resolved to its first single
+    base specification. See _The Variant Call Format Specification: VCFv4.4 and
+    BCFv2.2_ (2024-04-20) § 1.6.1.4 "Fixed fields: REF".
+
+[#268]: https://github.com/zaeleus/noodles/issues/268
+
 ## 0.57.0 - 2024-05-16
 
 ### Added